ocs refactoring, initial share handler
