+++ cmd/revad/svcs/httpsvcs/httpsvcs.go
@@ -28,13 +28,17 @@ import (
 // relative components before processing. head will never contain a slash and
 // tail will always be a rooted path without trailing slash.
 // see https://blog.merovius.de/2017/06/18/how-not-to-use-an-http-router.html
+// and https://gist.github.com/weatherglass/62bd8a704d4dfdc608fe5c5cb5a6980c#gistcomment-2161690 for the zero alloc code below
 func ShiftPath(p string) (head, tail string) {
-	p = path.Clean("/" + p)
-	i := strings.Index(p[1:], "/") + 1
-	if i <= 0 {
-		return p[1:], "/"
+	if p == "" {
+		return "", "/"
 	}
-	return p[1:i], p[i:]
+	p = strings.TrimPrefix(path.Clean(p), "/")
+	i := strings.Index(p, "/")
+	if i < 0 {
+		return p, "/"
+	}
+	return p[:i], p[i:]
 }
 
 // Service represents a HTTP service.
